requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pandas==2.3.3
numpy==2.4.2
MLB-StatsAPI==1.9.0
anthropic==0.83.0
streamlit==1.54.0
plotly==6.5.2
python-dotenv==1.2.1
requests==2.32.5
pydantic-settings
